added proof of concept for porting single maui test to xunit
