+++ src/System.Windows.Forms/tests/IntegrationTests/Common/ExternalTestHelpers.cs
