+++ src/System.Windows.Forms/tests/IntegrationTests/Common/TestHelpers.cs
